Change code layout
This commit changes the layout of the code to make it easier for newcomers to
start contributing to the project. The whole idea is to move everything that is
related to the gadget container image to the gadget-docker folder, so one can
realize very quickly what are those things for.
