+++ .gitignore
@@ -6,7 +6,7 @@
 *.dylib
 /inspektor-gadget
 /kubectl-gadget
-/gadget-ds/bin
+/gadget-container/bin
 
 # Test binary, build with `go test -c`
 *.test
