+++ gadget-container/gadget-from-local-bin.Dockerfile
