+++ gadget-container/gadget.Dockerfile
@@ -18,14 +18,15 @@ RUN set -ex; \
 	apt-get install -y --no-install-recommends \
 		ca-certificates curl
 
-COPY files/runc-hook-prestart.sh /bin/runc-hook-prestart.sh
-COPY files/runc-hook-poststop.sh /bin/runc-hook-poststop.sh
-COPY files/entrypoint.sh /entrypoint.sh
-COPY files/bcck8s /opt/bcck8s
+COPY entrypoint.sh /entrypoint.sh
+
+COPY ocihookgadget/runc-hook-prestart.sh /bin/runc-hook-prestart.sh
+COPY ocihookgadget/runc-hook-poststop.sh /bin/runc-hook-poststop.sh
+COPY bin/ocihookgadget /bin/ocihookgadget
 
 COPY bin/gadgettracermanager /bin/gadgettracermanager
-COPY bin/ocihookgadget /bin/ocihookgadget
+
+COPY gadgets/bcck8s /opt/bcck8s
 COPY bin/networkpolicyadvisor /bin/networkpolicyadvisor
 
 COPY --from=traceloop /bin/traceloop /bin/traceloop
-
