+++ gadget-container/gadgets/bcck8s/bcc-wrapper.sh
