+++ gadget-container/gadgettracermanager/main.go
