+++ pkg/gadgets/networkpolicy/networkpolicy_test.go
