+++ pkg/gadgets/networkpolicy/testdata/README.md
