+++ pkg/gadgets/networkpolicy/testdata/accept-from-pod.golden
