+++ pkg/gadgets/networkpolicy/testdata/accept-from-pod.input
