+++ pkg/gadgets/networkpolicy/testdata/connect-to-svc-duplicate.golden
