+++ pkg/gadgets/networkpolicy/testdata/connect-to-svc-namespace.golden
