+++ pkg/gadgets/networkpolicy/testdata/connect-to-svc.input
