+++ pkg/gadgets/networkpolicy/testdata/empty.input
