+++ pkg/gadgets/networkpolicy/testdata/multiple.golden
