+++ pkg/gadgets/networkpolicy/testdata/multiple.input
