+++ pkg/gadgets/networkpolicy/types/event.go
